Use correct workflows for license- and fossa-check (#12)
The "main" workflow was used erroneously. While this does perform the
license and fossa checks, it is for internal use by the .github repo,
and does not have the proper triggers for a reusable workflow call.
It is simple enough to split these two calls out and use their
respective workflows.

Signed-off-by: Eric Ball <eball@linuxfoundation.org>